fix some bug

+++ lib/H3SDK_BuildEnvironment.sh
@@ -89,14 +89,18 @@ function git_configure()
 
 function install_toolchain()
 {
-    mkdir -p $TOP_ROOT/OrangePiH3/.tmp_toolchain
-    cd $TOP_ROOT/OrangePiH3/.tmp_toolchain
-    curl -C - -o ./toolchain $toolchain
-    unzip $TOP_ROOT/OrangePiH3/.tmp_toolchain/toolchain
-    mkdir -p $TOP_ROOT/OrangePiH3/toolchain
-    mv $TOP_ROOT/OrangePiH3/.tmp_toolchain/OrangePiH3_toolchain-master $TOP_ROOT/OrangePiH3/toolchain/toolchain_tar
-    rm -rf $TOP_ROOT/OrangePiH3/.tmp_toolchain
-    cd -
+    if [ ! -d $TOP_ROOT/OrangePiH3/toolchain/arm-linux-gnueabi ]; then
+        mkdir -p $TOP_ROOT/OrangePiH3/.tmp_toolchain
+        cd $TOP_ROOT/OrangePiH3/.tmp_toolchain
+        curl -C - -o ./toolchain $toolchain
+        unzip $TOP_ROOT/OrangePiH3/.tmp_toolchain/toolchain
+        mkdir -p $TOP_ROOT/OrangePiH3/toolchain
+        mv $TOP_ROOT/OrangePiH3/.tmp_toolchain/OrangePiH3_toolchain-master $TOP_ROOT/OrangePiH3/toolchain/toolchain_tar
+        mv $TOP_ROOT/OrangePiH3/toolchain/toolchain_tar/OrangePiH3_toolchain-master/* $TOP_ROOT/OrangePiH3/toolchain/
+        rm -rf $TOP_ROOT/OrangePiH3/.tmp_toolchain
+        rm -rf $TOP_ROOT/OrangePiH3/toolchain/toolchain_tar
+        cd -
+    fi
 }
 
 git_configure
@@ -105,4 +109,7 @@ dirent_check
 install_toolchain
 end_op
 
-cd $TOP_ROOT
+whiptail --title "OrangePi Build System" --msgbox \
+ "`figlet OrangePi` Succeed to Create OrangePi Build System!        Path:$TOP_ROOT/OrangePiH53" \
+             15 50 0
+clear
